fix(): amount check
